backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.40.40",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-aws>=0.2.35",
    "langchain-community>=0.3.31",
    "mangum>=0.19.0",
    "openai>=1.109.1",
    "opensearch-py>=3.0.0",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests-aws4auth>=1.3.1",
    "uvicorn>=0.37.0",
]
